feat: always check if the child is exited in functions that return a feedback

+++ src/feedback.rs
@@ -23,7 +23,7 @@ pub enum Feedback {
     Variable(VariableValue),
     Stack(crate::stack::Stack),
     ProcessMap(Vec<proc_maps::MapRange>),
-    Exit(i16),
+    Exit(i32),
 }
 
 impl Display for Feedback {
